Update Flutter to use the latest dartdoc
Now that dart-lang/dartdoc#1236 is fixed.

+++ dev/bots/docs.sh
@@ -2,14 +2,34 @@
 set -e
 
 # Install dartdoc.
-# Versions after 0.9.7+1 suffer from https://github.com/dart-lang/dartdoc/issues/1236
-# so are we pinned to this old version until that bug is fixed.
-pub global activate dartdoc 0.9.7+1
+pub global activate dartdoc 0.9.9
 
 # This script generates a unified doc set, and creates
 # a custom index.html, placing everything into dev/docs/doc
 (cd dev/tools; pub get)
 FLUTTER_ROOT=$PWD dart dev/tools/dartdoc.dart
+
+# Smoke test the docs to make sure we have all the major kinds of things.
+
+if [[ ! -f dev/docs/doc/flutter/widgets/Widget/Widget.html ]]; then
+  echo 'Failed to find documentation for Widget class. Are the docs complete?'
+  exit 1
+fi
+
+if [[ ! -f dev/docs/doc/flutter/dart-io/File/File.html ]]; then
+  echo 'Failed to find documentation for File class. Are the docs complete?'
+  exit 1
+fi
+
+if [[ ! -f dev/docs/doc/flutter/dart-ui/Canvas/drawRect.html ]]; then
+  echo 'Failed to find documentation for Canvas.drawRect. Are the docs complete?'
+  exit 1
+fi
+
+if [[ ! -f dev/docs/doc/flutter/flutter_test/WidgetTester/pumpWidget.html ]]; then
+  echo 'Failed to find documentation for WidgetTester.pumpWidget. Are the docs complete?'
+  exit 1
+fi
 
 # Ensure google webmaster tools can verify our site.
 cp dev/docs/google2ed1af765c529f57.html dev/docs/doc
